Add RBT remove/fixup and tests.
Also share more code with BST. Does mean that RBT is aware of NIL
sentinel nodes in a couple of small places but it does significantly
reduce code duplication.

In test_rbt.js, add some utility functions to check the tree for
Red-Black properties. Add testRandom that inserts 15 nodes in random
order and removes them in a different order and checks the tree after
every insertion and removal for RBT property. Does this 10 times.